requirements-dev.txt
# Development dependencies for Alchemize
# Install with: pip install -r requirements-dev.txt

# Testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-xdist>=3.3.0  # For parallel test execution

# Code formatting and linting
black>=23.0.0
flake8>=6.0.0
flake8-docstrings>=1.7.0
flake8-import-order>=0.18.2
flake8-bugbear>=23.0.0
isort>=5.12.0

# Type checking
mypy>=1.5.0
types-requests>=2.31.0
types-redis>=4.6.0
types-python-dateutil>=2.8.19

# Security scanning
bandit>=1.7.5
safety>=2.3.0

# Development tools
pre-commit>=3.3.0
watchdog>=3.0.0  # For file watching during development
ipython>=8.14.0  # Enhanced REPL

# HTTP client for testing
httpx>=0.24.0

# Factory for creating test data
factory-boy>=3.3.0

# Faker for generating test data
faker>=19.0.0

# Coverage reporting
coverage[toml]>=7.2.0

# Performance profiling
py-spy>=0.3.14
memory-profiler>=0.61.0

# Load testing
locust>=2.15.0

# Debugging
pdb++>=0.10.3
icecream>=2.1.3  # Better print debugging

# Code complexity analysis
radon>=6.0.1

# Import sorting and formatting
autoflake>=2.2.0

# JSON schema validation for API testing
jsonschema>=4.18.0

# Time manipulation for testing
freezegun>=1.2.2

# Response mocking
responses>=0.23.0
requests-mock>=1.11.0

# Redis testing
fakeredis>=2.16.0

# File system testing
pyfakefs>=5.2.0

# API documentation testing
openapi-spec-validator>=0.6.0

# Performance monitoring
psutil>=5.9.0

# Database inspection
sqlalchemy-utils>=0.41.0

# Network security testing
requests-toolbelt>=1.0.0

# File format validation
python-magic>=0.4.27

# HTML parsing for content validation
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Color output for test results
colorama>=0.4.6

# Progress bars for long-running tests
tqdm>=4.65.0

# Dependency analysis
pipdeptree>=2.13.0

# Container testing
testcontainers>=3.7.0

# Property-based testing
hypothesis>=6.82.0

# Static analysis
pylint>=2.17.0

# Performance benchmarking
pytest-benchmark>=4.0.0

# Celery monitoring
flower>=2.0.0

# Metrics collection
prometheus-client>=0.17.0

# File system monitoring
watchdog>=3.0.0

# Network monitoring
netifaces>=0.11.0

# Exception monitoring
sentry-sdk>=1.29.0

# Notification services
slack-sdk>=3.21.0

# Data validation
cerberus>=1.3.4

# Vulnerability scanning
pip-audit>=2.6.0